// diantum-src/src/app/models/team-component.ts
export class TeamComponent {
    code: number;
    gender: string;
    color: string;
    type: number;
    name: string;
    objTotal: number;
    objPeriod: number;
    behav: any;
    performance: number;
    quality: number;
    changeCount: number;
    log: Array<string>;
    lastAction: number;
}